lex-mx-engine/pyproject.toml
[project]
name = "lex-mx-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.133.0",
    "greenlet>=3.3.2",
    "json-logic-qubit>=0.9.1",
    "jsonschema>=4.26.0",
    "litellm>=1.81.15",
    "playwright>=1.58.0",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "qdrant-client>=1.17.0",
    "sqlalchemy>=2.0.47",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "isort>=8.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.2",
]
